requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pyodbc==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
aiosmtplib==3.0.1
jinja2==3.1.2
python-dateutil==2.8.2
typing-extensions==4.8.0
email-validator==2.1.0